[gmx.py]: Add New Function `xopen`
Create a new function `xopen` that acts as a replacement for the
built-in `open` function but that can additionally read and write
compressed files.

Supported compression formats:

* gzip (.gz)
* bzip2 (.bz2)
* XZ/LZMA2 (.xz)
* LZMA (.lzma)

Replace all calls of the built-in `open` function by calls of `xopen`.

Update the allowed input types for the file name argument of functions
that open files from disk.