[submit_gmx_analyses_lintf2_ether.py]: Set Default of --every to 0
Change default of the --every option from 1 to 0.

The value of the --every option is parsed to the -dt option of the
submitted Gromacs tools.  In the Gromacs documentation
(https://manual.gromacs.org/documentation/2018-current/onlinehelp/gmx-trjconv.html)
it says about the -dt option:

> Only write/use frame when t MOD dt = first time (ps)

However, when -dt is set to 1 and the spacing between frames is less
than 1 ps, not every frame is processed, although t MOD 1 should be 0
for all (integer) values of t.  When setting -dt 0, every frame is
processed (although t MOD 0 should not be defined.  However, 0 is the
default value for -dt according to the docs).