[pr-labeler.yml]: Don't Run on Dependabot's PRs
Don't run pr-labeler on branches that were created by Dependabot,
because this removes the labels set by Dependabot.  `branches-ignore`
cannot be used, because this checks against the target branch of the
pull request, however we have to check against the head branch of the
pull request.  Solution: if statement, taken from
https://stackoverflow.com/questions/68388484/github-actions-exclude-pull-requests-from-a-branch