[submit_gmx_analyses_lintf2_ether.py]: Change Default of `--every`
Change the default value of the `--every` option from 0 to 1.

The value of the `--every` option is passed to the `-dt` option of the
submitted Gromacs tools that support this option.  The default value was
set to 0 in PR #74
(https://github.com/andthum/hpc_submit_scripts/pull/74),
because `gmx trjconv` did not process all frames when `-dt` was set to 1
and the spacing between trajectory frames was less than 1 ps.

However, it now turned out that the other Gromacs tools only process the
first frame when `-dt` is set to 0.  Therefore, we reset the default
value for `--every` (i.e. `-dt`) to 1.

In a following commit, we will replace the `-dt` option for `gmx
trjconv` with `-skip`, which should be more robust (see
https://www.mail-archive.com/gmx-users@gromacs.org/msg35903.html and
https://www.mail-archive.com/gmx-users@gromacs.org/msg06868.html).